mod16-cloudstorage/requirements.txt
flask
google-cloud-ndb
google-cloud-storage
